Move model filling methods into helper module.
